pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "expense-tracker"
version = "0.1.0"
description = "Self-hostable Telegram bot for tracking business trip expenses with Claude vision."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-telegram-bot>=21.5,<22",
    "sqlalchemy[asyncio]>=2.0.32",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.2",
    "aioboto3>=13.1.1",
    "anthropic>=0.39.0",
    "apscheduler>=3.10.4",
    "pydantic>=2.8.2",
    "pydantic-settings>=2.4.0",
    "openpyxl>=3.1.5",
    "reportlab>=4.2.2",
    "Pillow>=10.4.0",
    "structlog>=24.4.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.8",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
